Use int64_t for SQLite integers.
Use 64-bit integers with SQLite (when retrieving a column value as
well as when setting a parameter) instead of the "default" 32-bit int.

+++ database/target.cpp
@@ -19,22 +19,22 @@ TargetFinder::ProcessL1Targets (const HexCoord& centre,
 
   /* The query is actually about an L-infinity range, since that is easy
      to formulate in the database.  This certainly includes the L1 range.  */
-  stmt.Bind<int> (1, centre.GetX () - l1range);
-  stmt.Bind<int> (2, centre.GetX () + l1range);
-  stmt.Bind<int> (3, centre.GetY () - l1range);
-  stmt.Bind<int> (4, centre.GetY () + l1range);
+  stmt.Bind<int64_t> (1, centre.GetX () - l1range);
+  stmt.Bind<int64_t> (2, centre.GetX () + l1range);
+  stmt.Bind<int64_t> (3, centre.GetY () - l1range);
+  stmt.Bind<int64_t> (4, centre.GetY () + l1range);
 
   BindFactionParameter (stmt, 5, faction);
 
   auto res = stmt.Query ("targets");
   while (res.Step ())
     {
-      const HexCoord coord (res.Get<int> ("x"), res.Get<int> ("y"));
+      const HexCoord coord (res.Get<int64_t> ("x"), res.Get<int64_t> ("y"));
       if (HexCoord::DistanceL1 (centre, coord) > l1range)
         continue;
 
       proto::TargetId targetId;
-      targetId.set_id (res.Get<int> ("id"));
+      targetId.set_id (res.Get<int64_t> ("id"));
       targetId.set_type (proto::TargetId::TYPE_CHARACTER);
 
       cb (coord, targetId);
